# openapi/payment-api.yaml
openapi: 3.0.3
info:
  title: Payment Service API
  description: Balance check and payment processing for the online shop.
  version: 1.0.0
servers:
  - url: http://localhost:8081
    description: Local payment service
paths:
  /accounts/{accountId}:
    post:
      operationId: createAccount
      summary: Create an account with the default starting balance
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '201':
          description: Account created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BalanceResponse'
        '409':
          description: Account already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /accounts/{accountId}/balance:
    get:
      operationId: getBalance
      summary: Get current account balance
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Current balance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BalanceResponse'
        '404':
          description: Account not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /accounts/{accountId}/payment:
    post:
      operationId: makePayment
      summary: Withdraw an amount from the account balance
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentRequest'
      responses:
        '200':
          description: Payment processed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
        '404':
          description: Account not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Insufficient funds
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    BalanceResponse:
      type: object
      required:
        - balance
      properties:
        balance:
          type: integer
          format: int64
    PaymentRequest:
      type: object
      required:
        - amount
      properties:
        amount:
          type: integer
          format: int64
          minimum: 1
    PaymentResponse:
      type: object
      required:
        - balance
      properties:
        balance:
          type: integer
          format: int64
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          enum:
            - INSUFFICIENT_FUNDS
            - ACCOUNT_NOT_FOUND
            - ACCOUNT_ALREADY_EXISTS
        message:
          type: string